+++ src/editor/client/middlewares/localStorage.js
@@ -1,0 +1,1 @@
+// TODO use local storage to store last graph used