+++ src/editor/client/store/configureStore.js
@@ -1,4 +1,5 @@
 import { createStore, applyMiddleware, compose } from 'redux'
+import autorunMiddleware from '../middlewares/autorun'
 import canvasMiddleware from '../middlewares/canvas'
 import thunkMiddleware from 'redux-thunk'
 import rootReducer from '../reducers'
@@ -8,6 +9,7 @@ function configureStore (initialState) {
     rootReducer,
     initialState,
     compose(applyMiddleware(
+      autorunMiddleware,
       canvasMiddleware,
       thunkMiddleware
     ),