+++ gh-pages/changelog.md
@@ -37,6 +37,7 @@ Changelog format adheres to [Keep a Changelog](http://keepachangelog.com/)</sub>
 - flow view integration.
 - Custom nodes: toggle.
 - Load and save graph.
+- Autorun graph.
 
 [Unreleased]: https://github.com/fibo/{{ package.name }}/compare/v{{ package.version }}...HEAD
 {% for tag in tags offset:2 %}